Respect DUMP_BODY setting even when REJECT_MESSAGE is false
- It is often useful to have a copy of the original message when
malicious attachments are scrubbed before delivery. Dump the body
if DUMP_BODY is true even when the message is not rejected.

- When REJECT_MESSAGE is true, return as soon as a suspicious macro
is found instead of continung to scan the next attachment. This also
avoids dumping the body multiple times. And makes the code consistent
with the early return logic used when an attachment is found to be
in the hash database ('already parsed').

Signed-off-by: Selva Nair <selva.nair@gmail.com>